Fixed numerous oostore layer issues.

+++ charms/layers/oostore/reactive/oostore.py
@@ -1,14 +1,19 @@
+import os
 import shutil
 
 from charmhelpers.core import hookenv, host
 from charmhelpers.core.templating import render
 from charms.reactive import hook, when, when_not, is_state, set_state, remove_state
+from charms.reactive.bus import get_states
 
 
 @hook('install')
 def install():
     if is_state('oostore.available'):
         return
+    host.adduser('oostore', system_user=True)
+    host.add_group('oostore', system_group=True)
+    host.add_user_to_group('oostore', 'oostore')
     install_workload()
     set_state('oostore.available')
 
@@ -25,7 +30,10 @@ def upgrade():
 def install_workload():
     config = hookenv.config()
     deployment = config['deployment']
+    if not os.path.exists('/srv/oostore/%s/bin' % (deployment)):
+        os.makedirs('/srv/oostore/%s/bin' % (deployment), mode=0o755)
     shutil.copyfile('files/oostore', '/srv/oostore/%s/bin/oostore' % (deployment))
+    os.chmod('/srv/oostore/%s/bin/oostore' % (deployment), 0o755)
 
 
 @hook('config-changed')
@@ -39,7 +47,7 @@ def config_changed():
         if config.previous('https_port'):
             hookenv.close_port(config.previous('https_port'))
         hookenv.open_port(config['https_port'])
-    set_state('oostore.configured', config)
+    set_state('oostore.configured')
  
 
 @when('oostore.start')
@@ -56,49 +64,65 @@ def stop_oostore():
     remove_state('oostore.started')
 
 
-@when('oostore.configured', 'database.available')
-def setup(config, pg):
-    render(source="templates/upstart",
+@when('oostore.configured', 'database.connected', 'database.database.available')
+def setup(pg, _):
+    config = hookenv.config()
+    render(source="upstart",
         target="/etc/init/oostore.conf",
         owner="root",
         perms=0o644,
         context={
-            cfg: config,
-            db: pg,
+            'cfg': config,
+            'db': pg,
         })
 
-    cert_file = '/srv/oostore/%s/etc/cert.pem' % (config['deployment'])
+    deployment = config['deployment']
+    if not os.path.exists('/srv/oostore/%s/etc' % (deployment)):
+        os.makedirs('/srv/oostore/%s/etc' % (deployment), mode=0o755)
+
+    cert_file = '/srv/oostore/%s/etc/cert.pem' % (deployment)
     if config.get('cert'):
         with open(cert_file, "w") as fh:
             fh.write(config['cert'])
     else:
-        os.unlink(cert_file)
+        if os.path.exists(cert_file):
+            os.unlink(cert_file)
 
-    key_file = '/srv/oostore/%s/etc/key.pem' % (config['deployment'])
+    key_file = '/srv/oostore/%s/etc/key.pem' % (deployment)
     if config.get('key'):
         with open(key_file, "w") as fh:
             fh.write(config['key'])
         os.chmod(key_file, 0o600)
     else:
-        os.unlink(key_file)
+        if os.path.exists(cert_file):
+            os.unlink(key_file)
 
     set_state('oostore.start')
-    status_set('maintenance', 'Starting oostore')
+    hookenv.status_set('maintenance', 'Starting oostore')
 
 
+@when('oostore.available')
 @when_not('database.connected')
 def missing_db():
+    hookenv.log("%s" % (str(get_states())))
     remove_state('oostore.start')
     hookenv.status_set('blocked', 'Please add relation to postgresql')
 
 
 @when('database.connected')
-@when_not('database.available')
-def waiting_db():
+@when_not('database.database.available')
+def waiting_db(pg):
+    hookenv.log("%s" % (str(get_states())))
     remove_state('oostore.start')
     hookenv.status_set('waiting', 'Waiting for postgresql')
 
 
 @when('oostore.started')
 def oostore_started():
+    config = hookenv.config()
+    hookenv.open_port(config['http_port'])
+    if config.get('cert') and config.get('key'):
+        hookenv.open_port(config['https_port'])
+    
     hookenv.status_set('active', 'Ready')
+
